+++ src/main/java/btwg/api/block/blocks/BTWGGravelBlock.java
@@ -2,6 +2,7 @@ package btwg.api.block.blocks;
 
 import btw.block.BTWBlocks;
 import btw.block.blocks.FallingFullBlock;
+import btw.world.util.difficulty.DifficultyParam;
 import btwg.api.block.StoneType;
 import btwg.mod.BetterThanWorldGen;
 import net.minecraft.src.*;
@@ -61,6 +62,6 @@ public class BTWGGravelBlock extends FallingFullBlock {
 
     @Override
     public boolean isBreakableBarricade(World world, int i, int j, int k) {
-        return world.getDifficulty().canZombiesBreakBlocks();
+        return world.getDifficultyParameter(DifficultyParam.CanZombiesBreakBlocks.class);
     }
 }