+++ src/main/java/btwg/api/block/blocks/LooseStoneSlab.java
@@ -1,6 +1,7 @@
 package btwg.api.block.blocks;
 
 import btw.block.blocks.MortarReceiverSlabBlock;
+import btw.world.util.difficulty.DifficultyParam;
 import btwg.api.block.StoneType;
 import btwg.mod.BetterThanWorldGen;
 import net.fabricmc.api.EnvType;
@@ -64,7 +65,7 @@ public class LooseStoneSlab extends MortarReceiverSlabBlock {
 
     @Override
     public boolean isBreakableBarricade(World world, int i, int j, int k) {
-        return world.getDifficulty().canZombiesBreakBlocks();
+        return world.getDifficultyParameter(DifficultyParam.CanZombiesBreakBlocks.class);
     }
 
     //------ Client Side Functionality ------//